feat(tickets): packed trait-bucket lanes, seated round drain, unit-priced write budget, and the phase-end jackpot ticket tx split in two
Trait buckets (`lvlTraitEntry`) now hold eight uint32 lanes per slot naming positions in a
per-level append-only owner registry (`lvlEntryOwner`, slot 67), so materializing an entry costs a
lane share instead of a fresh 22k slot. The drain seats eight queue entries per round
(`FoilPackModule.drainRounds`), writes one whole word per quadrant, and resumes from a persisted
frontier plus the seated-index word (slot 68). Every sink registers the owner at queue time and the
owed word carries the position (uint80, bits 48..79). The write budget is priced in units bounded
by UNIT_GAS_BOUND with every step reserving its opcode-level worst case, so a drain call's work is a
pure function of state; `scripts/check-gasleft.sh` gates that.

The level-100 phase-end daily paid both 100-winner ticket legs off one word in one tx (16.1M with
up-front registration). The carryover leg is now its own advance stage
(`STAGE_JACKPOT_CARRYOVER_TICKETS`, 13), reached like every other stage: the coin+tickets stage
leaves Phase 1's packed budgets in place and keeps the lock, the next advance re-rolls the bonus
board from the same recorded word, pays the leg, zeroes the word and seals the day. Measured cold
at every winner cap: 8.99M then 7.16M.

Layout goldens re-captured; advance-call, RNG-window, RNG-taint, unchecked and write-owner
manifests updated; raw-store harnesses seed the owner registry; the mint-boost source pins follow
`_rollRemainder` into the storage base.

+++ test/fuzz/KeeperRouterOneCategory.t.sol
@@ -680,7 +680,7 @@ contract KeeperRouterOneCategory is DeployProtocol {
     ///      LR_INDEX-1 — the finalized index the relocated multi-index sweep reads.
     function _advanceLootboxRngIndexByOne() internal {
         uint256 packed = uint256(vm.load(address(game), bytes32(uint256(LOOTBOX_RNG_PACKED_SLOT))));
-        uint48 idx = uint48(packed & 0xFFFFFFFFFFFF);
+        uint80 idx = uint80(packed & 0xFFFFFFFFFFFF);
         packed = (packed & ~uint256(0xFFFFFFFFFFFF)) | uint256(idx + 1);
         vm.store(address(game), bytes32(uint256(LOOTBOX_RNG_PACKED_SLOT)), bytes32(packed));
     }
